Split sqlfile_test.go into reader,writer_test.go
