+++ pkg/sqlfs/reader_test.go
@@ -16,49 +16,14 @@ package sqlfs
 import (
 	"fmt"
 	"io"
-	"log"
 	"math/rand"
-	"os"
 	"strings"
-	"sync"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	_ "sqlflow.org/gohive"
 	"sqlflow.org/sqlflow/pkg/database"
 )
-
-const testDatabaseName = `sqlfs_test`
-
-var (
-	createSQLFSTestingDatabaseOnce sync.Once
-)
-
-func createSQLFSTestingDatabase() {
-	db := database.GetTestingDBSingleton()
-	stmt := fmt.Sprintf("CREATE DATABASE IF NOT EXISTS %s;", testDatabaseName)
-	if _, e := db.Exec(stmt); e != nil {
-		log.Fatalf("Cannot create sqlfs testing database %s: %v", testDatabaseName, e)
-	}
-}
-
-func TestWriterCreate(t *testing.T) {
-	createSQLFSTestingDatabaseOnce.Do(createSQLFSTestingDatabase)
-	db := database.GetTestingDBSingleton()
-	a := assert.New(t)
-
-	tbl := fmt.Sprintf("%s.unittest%d", testDatabaseName, rand.Int())
-	w, e := Create(db.DB, db.DriverName, tbl, database.GetSessionFromTestingDB())
-	a.NoError(e)
-	a.NotNil(w)
-	defer w.Close()
-
-	has, e1 := hasTable(db.DB, tbl)
-	a.NoError(e1)
-	a.True(has)
-
-	a.NoError(dropTable(db.DB, tbl))
-}
 
 func TestWriteAndRead(t *testing.T) {
 	a := assert.New(t)
@@ -117,20 +82,4 @@ func TestWriteAndRead(t *testing.T) {
 	a.NoError(r.Close())
 
 	a.NoError(dropTable(db.DB, tbl))
-}
-
-// assertNoError prints the error if there is any in TestMain, which
-// log doesn't work.
-func assertNoErr(e error) {
-	if e != nil {
-		log.Fatal(e)
-	}
-}
-
-func getEnv(key, fallback string) string {
-	value := os.Getenv(key)
-	if len(value) == 0 {
-		return fallback
-	}
-	return value
 }
